done d

+++ film.php
@@ -19,8 +19,34 @@
         </div>
 
         <div class="bar">
+            <div id="mySidenav" class="sidenav">
+                <a href="javascript:void(0)" class="closebtn" onclick="closeNav()">&times;</a>
+                <a href= "actor.php">actor</a>
+                <a href= "address.php">address</a>
+                <a href= "category.php">category</a>
+                <a href= "city.php">city</a>
+                <a href= "country.php">country</a>
+                <a href= "customer.php">customer</a>
+                <a href= "film_actor.php">film_actor</a>
+                <a href= "film_category.php">film_category</a>
+                <a href= "film_text.php">film_text</a>
+                <a href= "film.php">film</a>
+                <a href= "inventory.php">inventory</a>
+                <a href= "language.php">language</a>
+                <a href= "payment.php">payment</a>
+                <a href= "rental.php">rental</a>
+                <a href= "staff.php">staff</a>
+                <a href= "store.php">store</a>
+                <a href= "payment&customer.php">payment&customer</a>
+                <a href= "film&film_text&film_category&category&language.php">film&film_text&film_category&category&language</a>
+                <a href= "country&city&address.php">country&city&address</a>
+            </div>
+
+            <div class="menu">
+                <span style="font-size:25px;cursor:pointer;color:white; text-align:left; display:inline;" onclick="openNav()">&#9776; MENU</span>
+            </div>
             <form action= "" method= "POST" style= 'display: inline;'>
-                <input type ="text" name= "search" placeholder="SEARCH BY ID" style= "border-radius: 5px;">
+                <input type ="text" name= "search" placeholder="SEARCH BY FILM ID" style= "border-radius: 5px;">
             </form>
             <form action= "" method= "POST" style= 'display: inline;'>
                 <input type = "submit" name= "reset" value= "RESET" class="button"> 
@@ -94,37 +120,38 @@
                     <?php
                     if(isset($_POST['insert'])){
                         if(!empty($_POST['filmid'])&& !empty($_POST['releaseyear'])&& !empty($_POST['languageid'])&& !empty($_POST['rentalduration'])&& !empty($_POST['rentalrate'])&& !empty($_POST['length'])&& !empty($_POST['replacementcost'])&& !empty($_POST['rating'])&& !empty($_POST['specialfeatures'])){
-                        $filmid= $_POST['filmid'];
-                        $query= "SELECT film_id FROM film WHERE film_id = $filmid;";
-                        $result= mysqli_query($conn,$query);
-
-                            if(mysqli_num_rows($result)==0){
+                            if((is_numeric($_POST['filmid'])) && ($_POST['filmid'] > 0)){
                                 $filmid= $_POST['filmid'];
-                                $releaseyear= $_POST['releaseyear'];
-                                $languageid= $_POST['languageid'];
-                                $rentalduration= $_POST['rentalduration'];
-                                $rentalrate= $_POST['rentalrate'];
-                                $length= $_POST['length'];
-                                $replacementcost= $_POST['replacementcost'];
-                                $rating= $_POST['rating'];
-                                $specialfeatures= $_POST['specialfeatures'];
-                                $lastupdate= date('Y-m-d H:i:s');
-                                $insert = "INSERT INTO film VALUES('$filmid','$releaseyear','$languageid','$rentalduration','$rentalrate','$length','$replacementcost','$rating','$specialfeatures','$lastupdate');";
-                                $result = mysqli_query($conn,$insert);
-                                if ($result) {
-                                    echo '<script> alert("DATA INSERTED SUCCESSFULLY!")</script>';
+                                $query= "SELECT film_id FROM film WHERE film_id = $filmid;";
+                                $result= mysqli_query($conn,$query);
+
+                                if(mysqli_num_rows($result)==0){
+                                    $filmid= $_POST['filmid'];
+                                    $releaseyear= $_POST['releaseyear'];
+                                    $languageid= $_POST['languageid'];
+                                    $rentalduration= $_POST['rentalduration'];
+                                    $rentalrate= $_POST['rentalrate'];
+                                    $length= $_POST['length'];
+                                    $replacementcost= $_POST['replacementcost'];
+                                    $rating= strtoupper($_POST['rating']);
+                                    $specialfeatures= $_POST['specialfeatures'];
+                                    $lastupdate= date('Y-m-d H:i:s');
+                                    $insert = "INSERT INTO film VALUES('$filmid','$releaseyear','$languageid','$rentalduration','$rentalrate','$length','$replacementcost','$rating','$specialfeatures','$lastupdate');";
+                                    $result = mysqli_query($conn,$insert);
+                                    if ($result) {
+                                        echo '<script> alert("DATA INSERTED SUCCESSFULLY!")</script>';
+                                    }
+                                    else
+                                        echo '<script> alert("PREVIOUS INSERT FAILED! EITHER LANGUAGE ID ENTERED DOES NOT EXIST, CHECK THE LANGUAGE TABLE FOR AN EXISTING LANGUAGE ID OR CHECK IF THERE WERE MISTAKES MADE WHEN INSERTING")</script>';    
+                                    echo("<meta http-equiv='refresh' content='1'>");
+                                }else{
+                                    echo '<script> alert("PREVIOUS INSERT FAILED! FILM ID ENTERED ALREADY EXISTS")</script>';
                                 }
-                                else
-                                    echo '<script> alert("PREVIOUS INSERT FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN INSERTING")</script>';    
-                                echo("<meta http-equiv='refresh' content='1'>");
+                            }else{
+                                echo '<script> alert("PREVIOUS INSERT FAILED! INVALID FILM ID ENTERED")</script>';
                             }
-                            else{
-                                echo '<script> alert("PREVIOUS INSERT FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN INSERTING")</script>';
-                            }
-                                
-                        } 
-                        else{
-                            echo '<script> alert("PREVIOUS INSERT FAILED, PLEASE FILL ALL FIELDS!")</script>';
+                        }else{
+                            echo '<script> alert("PREVIOUS INSERT FAILED! PLEASE FILL ALL FIELDS")</script>';
                         }
                     }
                     
@@ -159,36 +186,39 @@
 
                     <?php
                     if(isset($_POST['update'])){
-                        if(!empty($_POST['filmid'])&& !empty($_POST['releaseyear'])&& !empty($_POST['languageid'])&& !empty($_POST['rentalduration'])&& !empty($_POST['rentalrate'])&& !empty($_POST['length'])){
-                            $filmid= $_POST['filmid'];
-                            $query= "SELECT film_id FROM film WHERE film_id = $filmid;";
-                            $result= mysqli_query($conn,$query);
-
-                            if(mysqli_num_rows($result)==1){
+                        if(!empty($_POST['filmid'])&& !empty($_POST['releaseyear'])&& !empty($_POST['languageid'])&& !empty($_POST['rentalduration'])&& !empty($_POST['rentalrate'])&& !empty($_POST['length'])&& !empty($_POST['replacementcost'])&& !empty($_POST['rating'])&& !empty($_POST['specialfeatures'])){
+                            if((is_numeric($_POST['filmid'])) && ($_POST['filmid'] > 0)){
                                 $filmid= $_POST['filmid'];
-                                $releaseyear= $_POST['releaseyear'];
-                                $languageid= $_POST['languageid'];
-                                $rentalduration= $_POST['rentalduration'];
-                                $rentalrate= $_POST['rentalrate'];
-                                $length= $_POST['length'];
-                                $replacementcost= $_POST['replacementcost'];
-                                $rating= $_POST['rating'];
-                                $specialfeatures= $_POST['specialfeatures'];
-                                $lastupdate= date('Y-m-d H:i:s');
-                                $update = "UPDATE film SET release_year= '$releaseyear', language_id= '$languageid',rental_duration='$rentalduration',rental_rate='$rentalrate',length='$length',replacement_cost='$replacementcost',rating='$rating',special_features='$specialfeatures',last_update='$lastupdate' WHERE film_id = $filmid;";
-                                $result = mysqli_query($conn,$update);
-                                if($result)
-                                    echo '<script> alert("DATA UPDATED SUCCESSFULLY!")</script>';
-                                else
-                                    echo '<script> alert("PREVIOUS UPDATE FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN UPDATING")</script>'; 
+                                $query= "SELECT film_id FROM film WHERE film_id = $filmid;";
+                                $result= mysqli_query($conn,$query);
+
+                                if(mysqli_num_rows($result)==1){
+                                    $filmid= $_POST['filmid'];
+                                    $releaseyear= $_POST['releaseyear'];
+                                    $languageid= $_POST['languageid'];
+                                    $rentalduration= $_POST['rentalduration'];
+                                    $rentalrate= $_POST['rentalrate'];
+                                    $length= $_POST['length'];
+                                    $replacementcost= $_POST['replacementcost'];
+                                    $rating= strtoupper($_POST['rating']);
+                                    $specialfeatures= $_POST['specialfeatures'];
+                                    $lastupdate= date('Y-m-d H:i:s');
+                                    $update = "UPDATE film SET release_year= '$releaseyear', language_id= '$languageid',rental_duration='$rentalduration',rental_rate='$rentalrate',length='$length',replacement_cost='$replacementcost',rating='$rating',special_features='$specialfeatures',last_update='$lastupdate' WHERE film_id = $filmid;";
+                                    $result = mysqli_query($conn,$update);
+                                    if($result)
+                                        echo '<script> alert("DATA UPDATED SUCCESSFULLY!")</script>';
+                                    else
+                                        echo '<script> alert("PREVIOUS UPDATE FAILED! EITHER LANGUAGE ID ENTERED DOEST NOT EXIST, CHECK THE LANGUAGE TABLE FOR AN EXISTING LANGUAGE ID OR CHECK IF THERE WERE MISTAKES MADE WHEN UPDATING")</script>'; 
                                 
-                                echo("<meta http-equiv='refresh' content='1'>");
+                                    echo("<meta http-equiv='refresh' content='1'>");
+                                }else{
+                                    echo '<script> alert("PREVIOUS UPDATE FAILED! FILM ID ENTERED DOES NOT EXIST")</script>';
+                                }
                             }else{
-                                echo '<script> alert("PREVIOUS UPDATE FAILED! CHECK IF THERE WERE MISTAKES MADE WHEN UPDATING")</script>';
+                                echo '<script> alert("PREVIOUS UPDATE FAILED! INVALID FILM ID ENTERED")</script>';
                             }
-                                
                         }else{
-                            echo '<script> alert("PREVIOUS UPDATE FAILED, PLEASE FILL ALL FIELDS!")</script>';
+                            echo '<script> alert("PREVIOUS UPDATE FAILED! PLEASE FILL ALL FIELDS")</script>';
                         }
                     }
                     ?>
@@ -210,26 +240,29 @@
             <?php
                 if(isset($_POST['delete'])){
                     if(!empty($_POST['filmid'])){
-                        $filmid= $_POST['filmid'];
-                        $query= "SELECT film_id FROM film WHERE film_id = $filmid;";
-                        $result= mysqli_query($conn,$query);
-
-                        if(mysqli_num_rows($result)==1){
+                        if((is_numeric($_POST['filmid'])) && ($_POST['filmid'] > 0)){
                             $filmid= $_POST['filmid'];
-                            $delete = "DELETE FROM film WHERE film_id= '$filmid'; ";
-                            $result = mysqli_query($conn,$delete); 
-                            if($result){
-                                echo '<script> alert("ROW DELETED SUCCESSFULLY!")</script>';
+                            $query= "SELECT film_id FROM film WHERE film_id = $filmid;";
+                            $result= mysqli_query($conn,$query);
+
+                            if(mysqli_num_rows($result)==1){
+                                $filmid= $_POST['filmid'];
+                                $delete = "DELETE FROM film WHERE film_id= '$filmid'; ";
+                                $result = mysqli_query($conn,$delete); 
+                                if($result){
+                                    echo '<script> alert("ROW DELETED SUCCESSFULLY!")</script>';
+                                }else{
+                                    echo '<script> alert("DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
+                                }
+                                echo("<meta http-equiv='refresh' content='1'>");
                             }else{
-                                echo '<script> alert("DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
+                                echo '<script> alert("PREVIOUS DELETE FAILED! FILM ID ENTERED DOES NOT EXIST")</script>';
                             }
-                            echo("<meta http-equiv='refresh' content='1'>");
                         }else{
-                            echo '<script> alert("PREVIOUS DELETE FAILED! YOU ARE NOT ALLOWED TO DELETE THIS ROW")</script>';
+                            echo '<script> alert("PREVIOUS DELETE FAILED! INVALID FILM ID ENTERED")</script>';
                         }
-                            
                     }else{
-                        echo '<script> alert("PREVIOUS DELETE FAILED, PLEASE FILL ALL FIELDS!")</script>';
+                        echo '<script> alert("PREVIOUS DELETE FAILED! PLEASE FILL ALL FIELDS")</script>';
                     }
                 }
             ?>
@@ -266,6 +299,14 @@
             function(){
                 document.querySelector('.delete').style.display= 'none';
             });
+            
+            function openNav() {
+                document.getElementById("mySidenav").style.width = "620px";
+            }
+
+            function closeNav() {
+                document.getElementById("mySidenav").style.width = "0";
+            }
         </script>
     </body>
 </html>
